cpu: Fix bug exposed by clang 18's -Woverloaded-virtual (#2127)
It looks like TAGE_SC_L's updatePathAndGlobalHistory wasn't updated with
the extra TAGEBase::BranchInfo *bi argument added in 09c7e0f1800.

Clang sees this as a mismatch between the number of args in the base
virtual function (with 6 args) and the overriden function (with 5 args)

The override fails to set `bi->modified = true` as the base
implementation updatePathAndGlobalHistory does.

To avoid this happening again with other compilers, add the `override`
specifier in the class definition.