misc,tests,gpu-compute: '-t' -> '-j' for Daily GPU Tests
If GPU tests need to recompile code due to the cache being out-of-date,
this job was doing so single-threaded when in all other cases we compile
gem5 with `-j $(nproc)` (4 threads on our self-hosted action runners
at the time of writing).

Likewise `-t` is used to specify the number of tests to run in
parallel. This therefore was running `nproc` (4) in parallel.
This is not recommended due to the limited memory of the runners.

Therefore this change moves from single threaded build with the
multi-processes used to run tests to multi-threaded building and a
single process used to run the tests serially.