python: Add OptionalParam to support std::optional. (#2252)
This commit introduces `OptionalParam` to generate `std::optional`
parameters. This allows gem5 developer to define an optional parameter
without using magic default numbers.

For example, if we want a parameter to overwrite sid. We used to have
two parameters for this:

```
sid = Param.UInt32(0, 'SID')
set_ssid = Param.Bool(False, 'Whether to set the SSID')
```

This caused invalid states to handle, for example what if sid is set but
set_ssid remains default False.

After this commit, we are able to solve this problem:

```
sid = OptionalParam.UInt32('SID')
```

On the C++ side, the parameter map to a `std::optional<uint32_t>`
member. We eliminate the undesired states.

Limitation:

OptionalParam only supports primitive types currently. We use nullptr to
express the emptiness for SimObjects.

Change-Id: I7e13231448862af8c03eb06e0c9e53e0a82b2efd