configs: Enable bootloader argument for RISCV simulations (#2256)
A recent patch (PR #2231) removed the bootloader argument from
`configs/example/riscv/fs_linux.py` due to the entry already being
present in `Options.py`. However, the entry is under an `if
buildEnv["USE_ARM_ISA"]` condition, making it unavailable for RISCV.
This patch moves the argument to a different position in the
`Options.py` file to enable it for RISCV simulations too.