arch-riscv: fix vcpyvs µops for vector indexed loads (#2307)
The number of `VCpyVsMicroInst` needed for vector indexed loads was
being calculated incorrectly by using the destination's EEW instead of
the source index's one. This could cause execution bugs in certain
situations (#2248).

This patch fixes the issue by using the source index's EEW when
calculating the number of µops needed.